requirements.txt
requests
parsel
pillow
pytesseract
flask
telegram-ext
pytesseract
